refactor: separación de algunos componentes reutilizables (combobox sidebar header)
